requirements.txt
amqp==5.3.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ffmpeg-python==0.2.0
flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
kombu==5.4.2
MarkupSafe==2.1.5
prompt-toolkit==3.0.48
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
typing-extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.0.6
zipp==3.20.2
